requirements.txt
kafka-python>=0.9.0
kazoo>=1.3
mysql-python>=1.2.5
python-statsd>=1.6.3
pyyaml>=3.11
numpy
https://s3-us-west-2.amazonaws.com/artifacts.numenta.org/numenta/nupic/releases/nupic-0.2.6-cp27-none-linux_x86_64.whl
scipy
statsmodels

